requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
discord==1.0.1
discord.py==1.6.0
idna==2.10
multidict==5.1.0
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
spotipy==2.17.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
youtube-dl==2021.3.14

